% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeMedoidsIndices}
\alias{computeMedoidsIndices}
\title{computeMedoidsIndices}
\usage{
computeMedoidsIndices(pMedoids, ref_series)
}
\arguments{
\item{pMedoids}{External pointer}

\item{ref_series}{reference series}
}
\value{
A map serie number -> medoid index
}
\description{
Compute medoids indices
}
