% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{computeSynchrones}
\alias{computeSynchrones}
\title{computeSynchrones}
\usage{
computeSynchrones(medoids, getRefSeries, nb_ref_curves, nb_series_per_chunk,
  ncores_clust = 1, verbose = FALSE, parll = TRUE)
}
\arguments{
\item{medoids}{big.matrix of medoids (curves of same length as initial series)}

\item{getRefSeries}{Function to retrieve initial series (e.g. in stage 2 after series
have been replaced by stage-1 medoids)}

\item{nb_ref_curves}{How many reference series? (This number is known at this stage)}

\item{nb_series_per_chunk}{(~Maximum) number of series in each group, inside a task}

\item{ncores_clust}{"OpenMP" number of parallel clusterings in one task}

\item{verbose}{Level of verbosity (0/FALSE for nothing or 1/TRUE for all; devel stage)}

\item{parll}{TRUE to fully parallelize; otherwise run sequentially (debug, comparison)}
}
\value{
A big.matrix of size K1 x L where L = data_length
}
\description{
Compute the synchrones curves (sum of clusters elements) from a matrix of medoids,
using L2 distances.
}
