% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{computeWerDists}
\alias{computeWerDists}
\title{computeWerDists}
\usage{
computeWerDists(synchrones, ncores_clust = 1, verbose = FALSE,
  parll = TRUE)
}
\arguments{
\item{synchrones}{A big.matrix of synchrones, in rows. The series have same length
as the series in the initial dataset}

\item{ncores_clust}{"OpenMP" number of parallel clusterings in one task}

\item{verbose}{Level of verbosity (0/FALSE for nothing or 1/TRUE for all; devel stage)}

\item{parll}{TRUE to fully parallelize; otherwise run sequentially (debug, comparison)}
}
\value{
A matrix of size K1 x K1
}
\description{
Compute the WER distances between the synchrones curves (in rows), which are
returned (e.g.) by \code{computeSynchrones()}
}
