% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{curvesToContribs}
\alias{curvesToContribs}
\title{curvesToContribs}
\usage{
curvesToContribs(series, wf, ctype)
}
\arguments{
\item{series}{Matrix of series (in rows), of size n x L}

\item{wf}{Wavelet transform filter; see ?wavelets::wt.filter}

\item{ctype}{Type of contribution: "relative" or "absolute" (or any prefix)}
}
\value{
A matrix of size n x log(L) containing contributions in rows
}
\description{
Compute the discrete wavelet coefficients for each series, and aggregate them in
energy contribution across scales as described in https://arxiv.org/abs/1101.4744v2
}
